优化接口，调整swagger，增加接口@Login登录权限

+++ proprietor/web/src/main/java/com/jsy/community/controller/BannerController.java
@@ -1,6 +1,7 @@
 package com.jsy.community.controller;
 
 
+import com.jsy.community.annotation.auth.Login;
 import com.jsy.community.api.IBannerService;
 import com.jsy.community.constant.Const;
 import com.jsy.community.entity.BannerEntity;
@@ -9,9 +10,7 @@ import com.jsy.community.qo.proprietor.BannerQO;
 import com.jsy.community.utils.ValidatorUtils;
 import com.jsy.community.vo.BannerVO;
 import com.jsy.community.vo.CommonResult;
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
-import io.swagger.annotations.ApiParam;
+import io.swagger.annotations.*;
 import org.apache.dubbo.config.annotation.DubboReference;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
@@ -19,13 +18,12 @@ import org.springframework.web.multipart.MultipartFile;
 import java.util.List;
 
 /**
- * <p>
- * banner轮播图 前端控制器
- * </p>
- *
- * @author jsy
- * @since 2020-11-16
- */
+* @Description: 轮播图控制器
+ * @Param:
+ * @Return:
+ * @Author: chq459799974
+ * @Date: 2020/11/17
+**/
 @Api(tags = "轮播图控制器")
 @RestController
 @RequestMapping("/banner")
@@ -35,13 +33,29 @@ public class BannerController {
 	private IBannerService iBannerService;
 	
 	/**
+	 * @Description: 轮播图列表查询
+	 * @Param: [bannerQO]
+	 * @Return: com.jsy.community.vo.CommonResult
+	 * @Author: chq459799974
+	 * @Date: 2020/11/16
+	 **/
+	@ApiOperation("【轮播图】列表查询")
+	@PostMapping("list")
+	public CommonResult list(@RequestBody BannerQO bannerQO){
+		ValidatorUtils.validateEntity(bannerQO, BannerQO.queryBannerValidatedGroup.class);
+		List<BannerVO> returnList = iBannerService.queryBannerList(bannerQO);
+		return CommonResult.ok(returnList);
+	}
+	
+	/**
 	* @Description: 轮播图 上传
 	 * @Param: [file, bannerEntity]
 	 * @Return: com.jsy.community.vo.CommonResult
 	 * @Author: chq459799974
 	 * @Date: 2020/11/16
 	**/
-	@ApiOperation("轮播图 上传")
+	@ApiOperation("【轮播图】上传")
+	@Login
 	@PostMapping("upload")
 	public CommonResult upload(MultipartFile[] file, BannerEntity bannerEntity){
 		if(file == null || file.length == 0){
@@ -59,30 +73,19 @@ public class BannerController {
 	}
 	
 	/**
-	* @Description: 轮播图列表查询
-	 * @Param: [bannerQO]
-	 * @Return: com.jsy.community.vo.CommonResult
-	 * @Author: chq459799974
-	 * @Date: 2020/11/16
-	**/
-	@ApiOperation("轮播图 列表查询")
-	@PostMapping("list")
-	public CommonResult list(@ApiParam(value = "需要删除的id【数组】") @RequestBody BannerQO bannerQO){
-		ValidatorUtils.validateEntity(bannerQO, BannerQO.queryBannerValidatedGroup.class);
-		List<BannerVO> returnList = iBannerService.queryBannerList(bannerQO);
-		return CommonResult.ok(returnList);
-	}
-	
-	/**
 	* @Description: 轮播图 批量删除
 	 * @Param: [bannerQO]
 	 * @Return: com.jsy.community.vo.CommonResult
 	 * @Author: chq459799974
 	 * @Date: 2020/11/16
 	**/
-	@ApiOperation("轮播图 批量删除")
+	@ApiOperation("【轮播图】批量删除")
+	@Login
 	@DeleteMapping("")
-	public CommonResult deleteBanner(@ApiParam(value = "需要删除的ids集合")@RequestBody Long[] ids){
+	public CommonResult deleteBanner(@RequestBody Long[] ids){
+		if(ids.length == 0){
+			return CommonResult.error(JSYError.REQUEST_PARAM.getCode(),JSYError.REQUEST_PARAM.getMessage());
+		}
 		boolean result = iBannerService.deleteBannerBatch(ids);
 		return result ? CommonResult.ok() : CommonResult.error(JSYError.INTERNAL.getCode(),JSYError.INTERNAL.getMessage());
 	}
